fix(seo): Refactor homepage to Server Component for metadata and schema optimization
- Renamed `app/(landing)/page.tsx` to `app/(landing)/landing-client.tsx` to isolate client-side logic.
- Created a new `app/(landing)/page.tsx` as a Server Component.
- Added explicit metadata export to the homepage using `pageMetadata.home`.
- Moved `HomePageJsonLd` injection from the shared `app/(landing)/layout.tsx` to the specific homepage file.
- This resolves the issue where the homepage was missing optimized titles/descriptions and ensures `WebSite` schema is only present on the homepage, not subpages.
