+++ app/(landing)/layout.tsx
@@ -2,7 +2,6 @@ import Navigation from '@/app/modern/components/navigation';
 import Footer from '@/app/modern/components/footer';
 import { Toaster } from '@/components/ui/toaster';
 import { ThemeProvider } from 'next-themes';
-import { HomePageJsonLd } from '@/components/seo/json-ld';
 
 export default function LandingLayout({
   children,
@@ -11,7 +10,6 @@ export default function LandingLayout({
 }) {
   return (
     <ThemeProvider attribute="class" defaultTheme="light" enableSystem={false}>
-      <HomePageJsonLd />
       <div className="min-h-screen flex flex-col">
         <Navigation />
         <main className="flex-1">
